added permissions

+++ Customer/views.py
@@ -1,13 +1,13 @@
-from django.shortcuts import render,redirect
-from . forms import CustomerModelForm
+from django.shortcuts import render,redirect, HttpResponse
+from . forms import CustomerModelForm,UserCreationModelForm
 from .models import Customer
 from django.contrib.auth import login, authenticate, logout
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+from django.contrib.auth.forms import  AuthenticationForm
 from django.contrib.auth.decorators import login_required
 from django.contrib import messages
+from django.contrib.auth.models import Group
 
 
-# Create your views here.
 def main_page(request):
     return render(request, './../templates/index.html')
 
@@ -15,14 +15,20 @@ def main_page(request):
 @login_required(login_url="sign_in")
 def customer_create_form(request):
     if request.method == 'POST':
-        form = CustomerModelForm(request.POST)
-        if form.is_valid():
-            form.save()
-            messages.success(request, "Customer has been created")
-            return redirect('list_customer')
+        if request.user.groups.filter(permissions__name='Can add customer'):
+            form = UserCreationModelForm(request.POST)
+            if form.is_valid():
+                user = form.save()
+                username = form.cleaned_data.get('username')
+                group = Group.objects.get(name='customer')
+                user.groups.add(group)
+                messages.success(request, "Customer has been created for" + username)
+                return redirect('list_customer')
+        else:
+            return HttpResponse('Permission not allowed')
     
     else:
-        form = CustomerModelForm()
+        form = UserCreationModelForm()
     
     context = {
         'form': form
@@ -42,13 +48,19 @@ def customer_list(request):
     }
     return render(request,'list_customer.html',context)
 
+
 @login_required(login_url="sign_in")
 def delete_customer(request,pk):
     delete_cus = Customer.objects.get(id=pk)
     if request.method == 'POST':
-        delete_cus.delete()
-        messages.success(request, "Customer has been deleted")
-        return redirect('list_customer')
+        if request.user.groups.filter(permissions__name='Can delete customer'):
+            delete_cus.delete()
+            messages.success(request, "Customer has been deleted")
+            return redirect('list_customer')
+        else :
+            return HttpResponse('you donot have permisssion')
+            
+            
     context = {
         'object': delete_cus
     }
@@ -59,11 +71,14 @@ def update_customer(request, pk):
     customer = Customer.objects.get(id=pk)
     form = CustomerModelForm(instance=customer)
     if request.method == "POST":
-        form = CustomerModelForm(request.POST, instance=customer)
-        if form.is_valid():
-            form.save()
-            messages.success(request, "Customer has been updated")
-            return redirect("list_customer")
+        if request.user.groups.filter(permissions__name='Can change customer'):
+            form = CustomerModelForm(request.POST, instance=customer)
+            if form.is_valid():
+                form.save()
+                messages.success(request, "Customer has been updated")
+                return redirect("list_customer")
+        else:
+            return HttpResponse('Permission not granted')
 
     context = {"form": form}
     return render(request, "customer_update_form.html", context)
@@ -73,9 +88,11 @@ def update_customer(request, pk):
 
 def signup_page(request):
     if request.method == "POST":
-        form = UserCreationForm(request.POST)
+        form = UserCreationModelForm(request.POST)
         if form.is_valid():
-            form.save()
+            user1=form.save()
+            group = Group.objects.get(name='customer')
+            user1.groups.add(group)
             username = form.cleaned_data.get("username")
             raw_password = form.cleaned_data.get("password")
             user = authenticate(username=username, password=raw_password)
@@ -83,7 +100,7 @@ def signup_page(request):
             messages.success(request, "Successfully Sign Up")
             return redirect("list_customer")
     else:
-        form = UserCreationForm()
+        form = UserCreationModelForm()
     context =  {"form": form}
     return render(request, "signup.html", context)
